# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "visionatrix-client",
    "version": "0.8.0.dev0"
  },
  "paths": {
...
    "/api/workers_info": {
      "get": {
        "summary": "Workers Info",
        "description": "Fetches details about workers including their system and device information.\nThis endpoint allows filtering of workers based on their last active status and can also\nretrieve information for a specific worker if a worker ID is provided.\nUseful for monitoring and managing worker resources in distributed computing environments.",
        "operationId": "workers_info_api_workers_info_get",
        "parameters": [
          {
            "name": "last_seen_interval",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "The time interval in seconds within which workers must have marked themselves active. If specified, only workers who have reported activity within this interval will be returned.",
              "default": 0,
              "title": "Last Seen Interval"
            },
            "description": "The time interval in seconds within which workers must have marked themselves active. If specified, only workers who have reported activity within this interval will be returned."
          },
          {
            "name": "worker_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "An optional worker ID to retrieve details for a specific worker.",
              "default": "",
              "title": "Worker Id"
            },
            "description": "An optional worker ID to retrieve details for a specific worker."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkerDetails"
                  },
                  "title": "Response Workers Info Api Workers Info Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/worker_tasks": {
      "post": {
        "summary": "Worker Tasks To Give Set",
        "description": "Sets the tasks that a worker can work on. An empty list indicates that all tasks are allowed.\nThe administrator can set `tasks_to_give` for all workers, users only for their own.",
        "operationId": "worker_tasks_to_give_set_api_worker_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_worker_tasks_to_give_set_api_worker_tasks_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Worker tasks set successfully"
          },
          "404": {
            "description": "Worker not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Can't find `worker_id` worker."
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_worker_tasks_to_give_set_api_worker_tasks_post": {
        "properties": {
          "worker_id": {
            "type": "string",
            "title": "Worker Id",
            "description": "ID of the worker"
          },
          "tasks_to_give": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tasks To Give",
            "description": "List of tasks the worker can handle"
          }
        },
        "type": "object",
        "required": [
          "worker_id",
          "tasks_to_give"
        ],
        "title": "Body_worker_tasks_to_give_set_api_worker_tasks_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkerDetails": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "The primary key for the worker record, automatically incremented."
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "The foreign key from the 'users' table, non-nullable."
          },
          "worker_id": {
            "type": "string",
            "title": "Worker Id",
            "description": "Uniq identifier for the worker, constructed from user_id, hostname, device name and device index."
          },
          "worker_version": {
            "type": "string",
            "title": "Worker Version",
            "description": "Version of the worker"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen",
            "description": "The timestamp of the worker's last activity, stored in UTC."
          },
          "tasks_to_give": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tasks To Give",
            "description": "Specifies tasks that the worker can execute. An empty list indicates that all tasks are allowed."
          },
          "os": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Os",
            "description": "Operating system type of the worker's machine, such as 'posix' or 'nt'."
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version",
            "description": "The version of Python running on the worker's machine."
          },
          "embedded_python": {
            "type": "boolean",
            "title": "Embedded Python",
            "description": "Indicates whether the Python environment is embedded (portable).",
            "default": false
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name",
            "description": "Name of the computing device."
          },
          "device_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Type",
            "description": "Type of the computing device, such as 'cuda' or 'cpu'."
          },
          "vram_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vram Total",
            "description": "Total VRAM available on the device in bytes."
          },
          "vram_free": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vram Free",
            "description": "Free VRAM available on the device in bytes."
          },
          "torch_vram_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Torch Vram Total",
            "description": "Total VRAM managed by PyTorch in bytes."
          },
          "torch_vram_free": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Torch Vram Free",
            "description": "Free VRAM managed by PyTorch that is currently unused."
          },
          "ram_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ram Total",
            "description": "Total RAM available on the worker in bytes."
          },
          "ram_free": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ram Free",
            "description": "Free RAM available on the worker in bytes."
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "worker_id",
          "worker_version",
          "last_seen",
          "tasks_to_give"
        ],
        "title": "WorkerDetails"
      },
...
    }
  }
}